+++ nds-client/src/main/java/br/com/abril/nds/service/impl/RecolhimentoServiceImpl.java
@@ -1,6 +1,7 @@
 package br.com.abril.nds.service.impl;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -437,7 +438,7 @@ public class RecolhimentoServiceImpl implements RecolhimentoService {
 		chamadaEncalheCota.setFechado(false);
 		chamadaEncalheCota.setCota(cota);
 		
-		BigDecimal qtdPrevista = BigDecimal.ZERO;
+		BigInteger qtdPrevista = BigInteger.ZERO;
 		
 		if (estoqueProdutoCota != null) {
