+++ nds-client/src/test/java/br/com/abril/nds/repository/impl/MovimentoRepositoryImplTest.java
@@ -1,6 +1,7 @@
 package br.com.abril.nds.repository.impl;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Calendar;
 import java.util.List;
 
@@ -56,13 +57,13 @@ public class MovimentoRepositoryImplTest extends AbstractRepositoryImplTest  {
 		tipoMovimentoFaltaEm = Fixture.tipoMovimentoFaltaEm();
 		tipoMovimentoFaltaEm = merge(tipoMovimentoFaltaEm);
 		
-		EstoqueProduto estoqueProdutoVeja = Fixture.estoqueProduto(produtoEdicaoVeja, BigDecimal.ZERO);
+		EstoqueProduto estoqueProdutoVeja = Fixture.estoqueProduto(produtoEdicaoVeja, BigInteger.ZERO);
 		save(estoqueProdutoVeja);
 		
 		MovimentoEstoque movimentoEstoque = 
 			Fixture.movimentoEstoque(
 				null, produtoEdicaoVeja, tipoMovimentoFaltaEm, usuarioJoao, estoqueProdutoVeja,
-				calendar.getTime(), BigDecimal.TEN, StatusAprovacao.PENDENTE, "motivo");
+				calendar.getTime(), BigInteger.TEN, StatusAprovacao.PENDENTE, "motivo");
 		save(movimentoEstoque);
 	}
